Allow server path to be configured
In Maven multi-module projects, the server path defaults to the path to the pom.xml. This means that a project with N maven sub-modules will have N servers show up in teamserver.

This change allows the user to specify a server path that we will then put on the jvm args to the agent. One complete test suite run will show up as one server in TeamServer.