Start on refactoring
Moving more stuff up to AbstractMojo so Install and Verify can be a little dumber and we can cut back on duplicated work.

This makes things a lot easier to test and helps eliminate the horrible static field hack that we were using to share data between mojos. The static fields persisted between test methods making tests deterministic which we'd like to avoid.